refactor(context-bag): replace orientation dispatch with summary pre-compute
Drop Ariadne's eager kickoff message on newly-attached bags. The handler now
warms the shared `context_summaries` cache and persists the initial render
snapshot, but posts no message. Ariadne stays silent until the user sends
their first turn, at which point the cache is already hot.

This clears the way for a composer-level "context attachment" UX where the
user can see, edit, and remove attached refs before sending — the first
response is to their actual question rather than a preamble.

- Rename: ContextBagOrientationHandler → ContextBagPrecomputeHandler
          createContextBagOrientationWorker → createContextBagPrecomputeWorker
          CONTEXT_BAG_ORIENT queue → CONTEXT_BAG_PRECOMPUTE
          ContextBagOrientJobData → ContextBagPrecomputeJobData
- Worker: strip withCompanionSession, persona/trace/createMessage/jobQueue
          deps, AI generateText call, checkForUnseenMessages catch-up, and
          orient:\${bagId} clientMessageId. Keeps resolveBagForStream (with
          skipIfAlreadyRendered) and persistSnapshot for idempotent reruns.
- Handler: drop personaId from the dispatched payload + the no-persona guard
           since there is no message author to resolve anymore.
- Intent:  remove DiscussThreadIntent.orientationUserPrompt (dead code per
           INV-38) plus the matching IntentConfig field and test.
- Docs:    update stale comments in streams handlers/service, resolve.ts,
           persona-agent-worker, api.ts, and message-actions.ts.

Tests: 974 backend unit (was 971) + 1369 frontend unit pass, typecheck +
lint clean across the monorepo. Handler tests stay at 5 paths (lost the
no-persona case); discuss-thread intent test loses 1 assertion. Manual
verification of the end-to-end flow comes in Phase 5 when the composer
path lands.
